+++ lib/glimmer/libui/window_proxy.rb
@@ -59,7 +59,7 @@ module Glimmer
       private
       
       def build_control
-        ::LibUI.init
+        self.class.init
         super.tap do
           handle_listener('on_closing') do
             destroy